requirements.txt
pandas>=2.2.0
requests>=2.32.0
beautifulsoup4>=4.12
lxml>=5.0
openpyxl>=3.1
yfinance>=0.2
matplotlib>=3.8
charset-normalizer>=3.3
python-dateutil>=2.9.0
python-dotenv>=1.0.0
shioaji
streamlit>=1.32.0
google-generativeai>=0.4.0
